Initial commit of spock examples. Contains both a gradle build file and an Idea project.
Both the gradle build file and the idea project use dependancies found in the lib folder instead of grabbing the files from a remote repository. This is to keep the configuration simple and to ensure that both the idea project and the gradle build use the same versions of the jars.

To run the HelloSpock test from the command line using gradle simply execute gradle test.
To run the HelloSpock test from Idea, create a configuration for HelloSpock and execute it.